Replace isGrounded Raycast with ContactFilter2D Add Dynamic Gravity
While raycast is supposedly conventional, it has the downside of not being strictly related to the player sprite. In this case, the raycast was projected from the center, and thus having the edge of a player collider touching the edge of a platform didn't reset the jump nor doublejump, despite the player being technically on a platform.

The code for dynamic gravity is also integrated but this involves further tweaking with how jumping is handled (velocity vs. Force).

contactFilter and dynamic gravity code source: https://github.com/Unity-Technologies/PhysicsExamples2D/tree/2022.2