Made corrections to air dash, other minor impulse adjustements.
The air dash now considers user input exiting the dash, otherwise continues in the same direction as the airdash without preventing midair directional control as with previous iteration.

The double jump does have potential to launch very high if timed correctly, but since it utilizes the physics engine as intended, I'm tempted to leave it be and simply adapt for it in the game design.